Introduce SingleSourceLogic superclass
